+++ src/xclim/core/locales.py
@@ -54,6 +54,7 @@ from copy import deepcopy
 from pathlib import Path
 
 from xclim.core.formatting import AttrFormatter, default_formatter
+from xclim.core.utils import CaseInsensitiveDict
 
 TRANSLATABLE_ATTRS = [
     "long_name",
@@ -131,7 +132,7 @@ def get_local_dict(locale: str | Sequence[str] | tuple[str, dict]) -> tuple[str,
         return locale, deepcopy(_LOCALES[locale])
 
     if isinstance(locale[1], dict):
-        trans = locale[1]
+        trans = CaseInsensitiveDict(locale[1])
     else:
         # Thus, a string pointing to a json file
         trans = read_locale_file(locale[1])
@@ -189,9 +190,9 @@ def get_local_attrs(
 
     local_attrs = {}
     for ind in reversed(indicator):
-        local_attrs = local_attrs | loc_dict.get(ind, {})
+        local_attrs = local_attrs | loc_dict.get(ind.lower(), {})
         if var_name:
-            local_attrs = local_attrs | loc_dict.get(f"{ind}.{var_name}", {})
+            local_attrs = local_attrs | loc_dict.get(f"{ind}.{var_name}".lower(), {})
 
     attrs = {}
     if not local_attrs:
@@ -266,14 +267,13 @@ def read_locale_file(filename, module: str | None = None, encoding: str = "UTF8"
     Returns
     -------
     dict
-        The locale dictionary.
+        The locale dictionary. All entries are lowercase.
     """
-    locdict: dict[str, dict]
     with open(filename, encoding=encoding) as f:
-        locdict = json.load(f)
+        data = json.load(f)
 
-    if module is not None:
-        locdict = {(k if k == "attrs_mapping" else f"{module}.{k}"): v for k, v in locdict.items()}
+    modstr = f"{module}." if module is not None else ""
+    locdict = CaseInsensitiveDict({(k if k == "attrs_mapping" else f"{modstr}{k}"): v for k, v in data.items()})
     return locdict
 
 
@@ -298,7 +298,7 @@ def load_locale(locdata: str | Path | dict[str, dict], locale: str) -> None:
         _LOCALES[locale] = locdata
 
 
-def generate_local_dict(locale: str, init_english: bool = False) -> dict:
+def generate_local_dict(locale: str, init_english: bool = False) -> CaseInsensitiveDict:
     """
     Generate a dictionary with keys for each indicator and translatable attributes.
 
@@ -322,32 +322,35 @@ def generate_local_dict(locale: str, init_english: bool = False) -> dict:
             if ind_name != "attrs_mapping" and ind_name not in registry:
                 attrs.pop(ind_name)
     else:
-        attrs = {}
+        attrs = CaseInsensitiveDict()
 
     attrs_mapping = attrs.setdefault("attrs_mapping", {})
     attrs_mapping.setdefault("modifiers", [""])
     for key, value in default_formatter.mapping.items():
         attrs_mapping.setdefault(key, [value[0]])
 
-    eng_attr = ""
     for ind_name, indicator in registry.items():
         ind_attrs = attrs.setdefault(ind_name, {})
-        for translatable_attr in set(TRANSLATABLE_ATTRS).difference(set(indicator._cf_names)):
+        for prop in indicator._translatable_props:
             if init_english:
-                eng_attr = getattr(indicator, translatable_attr)
-                if not isinstance(eng_attr, str):
-                    eng_attr = ""
-            ind_attrs.setdefault(f"{translatable_attr}", eng_attr)
+                eng = getattr(indicator, prop)
+                if not isinstance(eng, str):
+                    eng = ""
+            else:
+                eng = ""
+            ind_attrs.setdefault(prop, eng)
 
-        for cf_attrs in indicator.cf_attrs:
+        for atts in indicator.attrs:
             # In the case of single output, put var attrs in main dict
-            if len(indicator.cf_attrs) > 1:
-                ind_attrs = attrs.setdefault(f"{ind_name}.{cf_attrs['var_name']}", {})
+            if len(indicator.attrs) > 1:
+                ind_attrs = attrs.setdefault(f"{ind_name}.{atts.var_name}", {})
 
-            for translatable_attr in set(TRANSLATABLE_ATTRS).intersection(set(indicator._cf_names)):
+            for attr in indicator._translatable_attrs:
                 if init_english:
-                    eng_attr = cf_attrs.get(translatable_attr)
-                    if not isinstance(eng_attr, str):
-                        eng_attr = ""
-                ind_attrs.setdefault(f"{translatable_attr}", eng_attr)
+                    eng = atts.get(attr)
+                    if not isinstance(eng, str):
+                        eng = ""
+                else:
+                    eng = ""
+                ind_attrs.setdefault(attr, eng)
     return attrs